allow to force memory of localstorage on storage
