+++ packages/sdk/src/base/storage/base.ts
@@ -1,8 +1,12 @@
-import { IStorage, StorageOptions } from '../../storage';
+import { IStorage } from '../../storage';
 
-export type { StorageOptions };
+export type StorageOptions = {
+	prefix?: string;
+};
 
 export abstract class BaseStorage extends IStorage {
+	protected prefix: string;
+
 	get auth_token(): string | null {
 		return this.get('auth_token');
 	}
@@ -46,4 +50,10 @@ export abstract class BaseStorage extends IStorage {
 	abstract get(key: string): string | null;
 	abstract set(key: string, value: string): string;
 	abstract delete(key: string): string | null;
+
+	constructor(options?: StorageOptions) {
+		super();
+
+		this.prefix = options?.prefix ?? '';
+	}
 }
